Make changes based on PR feedback

+++ src/generators.cpp
@@ -500,8 +500,8 @@ void Generator::SetRuntimeOption(const char* key, const char* value) {
   state_->SetRunOption(key, value);
 }
 
-int32_t Generator::TokenCount() const {
-  return static_cast<int32_t>(search_->GetSequenceLength());
+size_t Generator::TokenCount() const {
+  return static_cast<size_t>(search_->GetSequenceLength());
 }
 
 bool Generator::IsDone() {
